+++ lib/screens/login_screens/change_email_screen.dart
@@ -1,8 +1,6 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:mgs_app2/services/firebase/auth.dart';
-import 'package:mgs_app2/services/firebase/exceptions_translator.dart';
 import 'package:mgs_app2/utilities/app_config.dart';
 import 'package:mgs_app2/widgets/button.dart';
 import 'package:mgs_app2/widgets/snackbar.dart';
