+++ lib/screens/login_screens/login_screen.dart
@@ -1,6 +1,6 @@
 import 'package:mgs_app2/screens/login_screens/forgot_password_screen.dart';
-import 'package:mgs_app2/screens/login_screens/registration_controller.dart';
-import 'package:mgs_app2/screens/login_screens/registration_screen1.dart';
+import 'package:mgs_app2/screens/registration_screens/registration_controller.dart';
+import 'package:mgs_app2/screens/registration_screens/registration_screen1.dart';
 import 'package:flutter/material.dart';
 import 'package:mgs_app2/utilities/app_config.dart';
 import 'package:mgs_app2/widgets/font.dart';
