+++ lib/screens/registration_screens/registration_screen1.dart
@@ -1,8 +1,8 @@
 import 'package:mgs_app2/models/user_model.dart';
 import 'package:flutter/material.dart';
-import 'package:mgs_app2/screens/login_screens/optional_registration_screen1.dart';
-import 'package:mgs_app2/screens/login_screens/registration_controller.dart';
-import 'package:mgs_app2/screens/login_screens/registration_screen2.dart';
+import 'package:mgs_app2/screens/registration_screens/optional_registration_screen1.dart';
+import 'package:mgs_app2/screens/registration_screens/registration_controller.dart';
+import 'package:mgs_app2/screens/registration_screens/registration_screen2.dart';
 import 'package:mgs_app2/utilities/app_config.dart';
 import 'package:mgs_app2/widgets/back_button_app_bar.dart';
 import 'package:mgs_app2/widgets/font.dart';
